+++ src/Gateway/Winterhaven.Gateway.Infrastructure/Options/Client/ClientOptions.cs
@@ -1,0 +1,14 @@
+﻿using System.ComponentModel.DataAnnotations;
+using System.Diagnostics.CodeAnalysis;
+
+namespace Winterhaven.Gateway.Infrastructure.Options.Client;
+
+[ExcludeFromCodeCoverage(Justification = "No Logic")]
+internal sealed class ClientOptions
+{
+    [Required]
+    public required string ApiKey { get; init; }
+
+    [Required]
+    public required string BaseUrl { get; init; }
+}