+++ src/Gateway/Winterhaven.Gateway.Presentation/Services/Targets/IJsonRpcTargetRegistrar.cs
@@ -1,0 +1,8 @@
+﻿using StreamJsonRpc;
+
+namespace Winterhaven.Gateway.Presentation.Services.Targets;
+
+internal interface IJsonRpcTargetRegistrar
+{
+    public void RegisterTargets(JsonRpc rpc);
+}